requirements.txt
python-dotenv==1.2.2
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11
pgvector==0.4.2
discord.py==2.7.1
# jupyter==1.1.1
pandas==2.3.3
# matplotlib==3.10.8
# seaborn==0.13.2
langgraph==1.1.10
langchain-google-genai==4.2.1
langchain-groq==1.1.2
uvicorn==0.42.0
fastapi==0.135.2
# streamlit==1.55.0
neo4j==6.1.0
langchain-ollama==1.0.1
lightrag-hku==1.4.16
rank-bm25==0.2.2
